+++ CryptoBotCore/Models/WalletBalances.cs
@@ -8,11 +8,10 @@ namespace CryptoBotCore.Models
     public class WalletBalances
     {
 
-        public WalletBalances(string currency, Double available)
+        public WalletBalances(string currency, decimal available)
         {
             this.currency = currency;
             this.available = available;
-  
         }
 
         public string currency { get; set; }
@@ -20,6 +19,6 @@ namespace CryptoBotCore.Models
         /// <summary>
         /// How much X there is in this wallet that is available to trade
         /// </summary>
-        public double available { get; set; }
+        public decimal available { get; set; }
     }
 }
